Renamed files - Substituted - with _ as guidelines specify - Change direftories to follow same standard than rest of the app
