+++ lib/src/model/profile.dart
@@ -1,7 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-
-import '../utils/exceptions/user_already_exists.dart';
+import 'package:queasy/utils/exceptions/user_already_exists.dart';
 
 String? getCurrentUserID() {
   if (FirebaseAuth.instance.currentUser != null) {
@@ -101,9 +100,9 @@ class Profile {
   /// returns true if successful
   /// throws UserAlreadyExistsException if the user with the same username already exists in the database
   /// [firestore] database instance
-  bool registerUser(FirebaseFirestore firestore){
+  bool registerUser(FirebaseFirestore firestore) {
     try {
-      // doesnt work for now 
+      // doesnt work for now
       // firestore
       //     .collection('users')
       //     .doc(this.username)
@@ -119,9 +118,6 @@ class Profile {
       return false;
     }
   }
-
-
-
 
   /// Increment the score of the user in the firebase by the score achieved in the current quiz
   /// [username] The username of the user
