+++ lib/src/view/login/login_view.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:queasy/src/view/widgets/rounded-button.dart';
-import '../../services/auth.dart';
+import '../../../services/auth.dart';
 
 class LogInView extends StatefulWidget {
   const LogInView({Key? key}) : super(key: key);
