+++ lib/src/view/registration/register_desktop.dart
