+++ lib/utils/exceptions/user_already_exists.dart
